+++ packages/tiny-rpc/src/index.ts
@@ -1,10 +1,6 @@
 export * from "./core";
 export * from "./adapter-http";
 export * from "./adapter-message-port";
-export * from "./adapter-message-port-node";
 export * from "./validation";
-export {
-  TwoWaySseClient,
-  TwoWaySseClientProxy,
-  createTwoWaySseHandler,
-} from "./server-sent-event";
+export * from "./message-port/nodejs";
+export * from "./message-port/server-sent-event";
